First version able to render 1999 data as well as 2010 and 2018 data.
Ability to select 'comparative' metrics commented out, at least for now.: no support in the logic for 2018-1999 or 2010-1999 comparisons (yet). Wireframe + count data includes segments present in 1999 but not in 2010, and all segments are tagged with an indication of years in which they exist. GeoJSON inlucdes linework for 1999 as well as 2010/2018. AWDT data values for 1999 need to be checked: there's at least one apparent error here: R12209 AWDT in 1999.