Corrected errors in transcribing 1999 AWDT data values for several NB and SB segments.
There are still some issues TBD re: the 1999 AWDT data: (1) a copule of discontinuties between the value in one 'sheet' in Tom Lisco's stick diagrams and its neighbor, and (2) the fact that all '1999' AWDT values for Route 3 south are, in  fact, for 1997. Bill Kuttner will 'scale' these to 1999 values, but this has not yet been completed.